Force tmux to use the current path when opening a window
In a recent version of tmux the default behavior changed for opening
windows and panes. When opening a window or pane in tmux it now uses
the current path of the environment tmux was opened in. It used top open
it to the current path of the currently selected window or pane. This
commit restores the behavior of opening a new window or pane using the
current path of the current tmux pane or window you are in.