Add script to turn off symlinks
This is to begin the great migration to ansible. It allows for the files
in the repository to move around and be removed without impacting the
current environment in which this is being developed.

I know I should be using a test environment, but really I'm to lazy for
that right now.